tests - DeleteCategory command

+++ Expenses/Expenses.Api/src/Commands/DeleteCategoryCommandHandler.cs
@@ -11,19 +11,23 @@ namespace Expenses.Api.Commands {
         }
 
         public override async Task<bool> Handle(DeleteCategoryCommand command, CancellationToken ct) {
-            Category defaultCategory = await Repository.GetCategoryByIdAsync(Constants.DefaultCategoryId, ct);
+            if(command.CategoryId == Constants.DefaultCategoryId)
+                throw new DomainException(
+                    cause: DomainExceptionCause.DefaultCategoryUpdateOrDelete,
+                    message: "The default category cannot be deleted"
+                );
 
             Category category = await Repository.EnsureCategoryByIdAsync(command.CategoryId, ct);
+            Category defaultCategory = await Repository.GetCategoryByIdAsync(Constants.DefaultCategoryId, ct);
 
-            await Repository.LoadExpensesAsync(defaultCategory, ct);
             await Repository.LoadExpensesAsync(category, ct);
+            await Repository.LoadExpensesAsync(defaultCategory, ct);
 
             category.MoveExpenses(defaultCategory);
 
             Repository.RemoveCategory(category);
 
             await Repository.UnitOfWork.SaveAsync(ct);
-
 
             return true;
         }
